fix: route segment storybook 세팅 변경
