+++ src/styles/globalStyles.css.ts
@@ -1,5 +1,4 @@
 import { globalStyle } from "@vanilla-extract/css";
-import { theme } from "./themes.css";
 
 globalStyle("*, *::before, *::after", {
   boxSizing: "border-box",
@@ -8,7 +7,6 @@ globalStyle("*, *::before, *::after", {
   fontSize: process.env.STORYBOOK === "true" ? "inherit" : "62.5%",
   margin: "0",
   padding: 0,
-  backgroundColor: theme.color.bg,
 });
 
 globalStyle("html", {
